Editor: Make line numbers exact for files beyond two billion lines
Line numbers are now Long end-to-end: TextPosition.line, totalLines,
line-addressed buffer/engine APIs (getTextForLine, getTextForRange,
findOffset, goToLine, updateVisibleRange), search match lines, the
workspace state, and the persisted cursorLine/scrollToLine session
arguments (old Int payloads still restore). Line counting no longer
truncates through Int, so line-addressed navigation (go-to-line,
search jumps, cursor restore) stays exact at any depth.

Narrowing to Int happens only where a Compose/framework API forces it
(LazyColumn item counts, plural quantities), always saturating via
toIntSaturated() - lines past Int.MAX_VALUE are not scrollable in the
current UI but the engine addresses them correctly. Viewport caches
re-key to Long. Columns intentionally stay Int (Kotlin String and
Compose text layout are hard Int ceilings).

Adds a synthetic 2.29-billion-break document test pinning exact
lineStartOffset/lineOfOffset past Int.MAX_VALUE, and a session-restore
compat test for old Int line payloads.

+++ app-workspace-editor/src/main/java/eu/darken/butler/editor/core/engine/TextPosition.kt
@@ -1,9 +1,14 @@
 package eu.darken.butler.editor.core.engine
 
+/**
+ * [line] is Long end-to-end so line addressing stays exact on documents beyond 2^31 lines;
+ * narrowing to Int happens only where a Compose/framework API forces it (always saturating).
+ * [column] stays Int: it indexes a Kotlin String / Compose text layout, a hard Int ceiling.
+ */
 data class TextPosition(
     val offset: Long,
-    val line: Int,
-    val column: Int
+    val line: Long,
+    val column: Int,
 ) {
     companion object {
         val ZERO = TextPosition(0, 0, 0)
